Refactor AutocompleteRepo into SearchRepo and AutocompleteModule into SearchModule.
- https://github.com/alphagov/govuk-mobile-android-app/pull/170#discussion_r1923507317

- https://github.com/alphagov/govuk-mobile-android-app/pull/170#discussion_r1923508475
