+++ feature/search/src/main/kotlin/uk/govuk/app/search/data/SearchRepo.kt
@@ -3,7 +3,9 @@ package uk.govuk.app.search.data
 import uk.govuk.app.networking.domain.DeviceOfflineException
 import uk.govuk.app.networking.domain.ServiceNotRespondingException
 import uk.govuk.app.search.data.local.SearchLocalDataSource
+import uk.govuk.app.search.data.remote.AutocompleteApi
 import uk.govuk.app.search.data.remote.SearchApi
+import uk.govuk.app.search.data.remote.model.AutocompleteResponse
 import uk.govuk.app.search.data.remote.model.SearchResponse
 import uk.govuk.app.search.domain.SearchConfig
 import javax.inject.Inject
@@ -12,6 +14,7 @@ import javax.inject.Singleton
 @Singleton
 internal class SearchRepo @Inject constructor(
     private val searchApi: SearchApi,
+    private val autocompleteApi: AutocompleteApi,
     private val localDataSource: SearchLocalDataSource
 ) {
 
@@ -35,6 +38,19 @@ internal class SearchRepo @Inject constructor(
         return try {
             val response = searchApi.getSearchResults(searchTerm, count)
             Result.success(response)
+        } catch (_: java.net.UnknownHostException) {
+            Result.failure(DeviceOfflineException())
+        } catch (_: retrofit2.HttpException) {
+            Result.failure(ServiceNotRespondingException())
+        } catch (e: Exception) {
+            Result.failure(e)
+        }
+    }
+
+    suspend fun performLookup(searchTerm: String): Result<AutocompleteResponse> {
+        return try {
+            val response = autocompleteApi.getSuggestions(searchTerm)
+            return Result.success(response)
         } catch (_: java.net.UnknownHostException) {
             Result.failure(DeviceOfflineException())
         } catch (_: retrofit2.HttpException) {
